Reconocimiento de texto manuescrito a voz/requirements.txt
streamlit
opencv-python
numpy
easyocr
torch
transformers
Pillow
scipy
openai